requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cms==0.0.2
cycler==0.10.0
Django==3.0.7
django-crispy-forms==1.9.1
django-mathfilters==1.0.0
djangorestframework==3.11.0
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.0
nba-api==1.1.8
numpy==1.18.5
pandas==1.0.5
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
